Add Signal to interrupt child exiting * Also kill all forks just before exit

+++ childHandler.c
@@ -6,18 +6,28 @@ int numbgChilds=0;
 
 process bgProcess[50];
 
-int killChilds(){
-    int i, status;
-    for(i=0;i<numbgChilds;i++){
-        printf("%d \n", waitpid(bgProcess[i].id, &status, 0));
-        fprintf(stderr, "%s exited successfully pid: %d ", bgProcess[i].name, bgProcess[i].id);
+int getiofid(int id){
+    int i;
+    for(i=0;i<numbgChilds;i++)
+        if(bgProcess[i].id==id)
+            return i;
+}
+
+void killChilds(int id){
+    int status;
+    int pid=waitpid(-1, &status, WNOHANG);
+    if(pid>0){
+        printf("[] %s pid %d exited successfully\n", bgProcess[getiofid(pid)].name, pid);
+        bgProcess[getiofid(pid)].inbg=1;
     }
-    numbgChilds=0;
 }
 
 int killallChilds(){
     int i;
     for(i=0;i<numbgChilds;i++){
-        kill(bgProcess[i].id, SIGKILL);
+        if(!bgProcess[i].inbg){
+            kill(bgProcess[i].id, SIGKILL);
+            printf("[] %s Stopped pid: %d\n", bgProcess[i].name, bgProcess[i].id);
+        }
     }    
 }
